Refactor ConfirmationMeter logic
